+++ projects/packages/backup/src/js/selectors/site-backup.js
@@ -6,11 +6,13 @@ const siteBackupSelectors = {
 	getDaysOfBackupsAllowed: state => state.siteBackupSize.daysOfBackupsAllowed ?? null,
 	getDaysOfBackupsSaved: state => state.siteBackupSize.daysOfBackupsSaved ?? null,
 	getBackupRetentionDays: state => state.siteBackupSize.retentionDays ?? null,
+	hasBackupSizeLoaded: state => state.siteBackupSize.loaded,
 
 	// Policies
 	isFetchingBackupPolicies: state => state.siteBackupPolicies.isFetching ?? null,
 	getBackupStorageLimit: state => state.siteBackupPolicies.storageLimitBytes ?? null,
 	getActivityLogLimitDays: state => state.siteBackupPolicies.activityLogLimitDays ?? null,
+	hasBackupPoliciesLoaded: state => state.siteBackupPolicies.loaded,
 
 	// Storage
 	getStorageUsageLevel: state => state.siteBackupStorage.usageLevel ?? null,