+++ projects/packages/videopress/src/client/block-editor/blocks/video/components/videopress-uploader/index.native.js
@@ -3,14 +3,14 @@
  */
 import { BlockIcon, MediaPlaceholder } from '@wordpress/block-editor';
 import { useDispatch } from '@wordpress/data';
-import { useCallback, useState } from '@wordpress/element';
+import { useCallback, useEffect, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { store as noticesStore } from '@wordpress/notices';
-import { isURL, getProtocol } from '@wordpress/url';
 /**
  * Internal dependencies
  */
 import { buildVideoPressURL } from '../../../../../lib/url';
+import isLocalFile from '../../../../utils/is-local-file.native';
 import { VIDEOPRESS_VIDEO_ALLOWED_MEDIA_TYPES } from '../../constants';
 import { title } from '../../index';
 import { VideoPressIcon } from '../icons';
@@ -19,13 +19,32 @@ import './style.scss';
 
 const VideoPressUploader = ( {
 	autoOpenMediaUpload,
+	fileToUpload,
 	handleDoneUpload,
 	isInteractionDisabled,
+	isReplacing,
 	onFocus,
+	onReplaceCancel,
+	onStartUpload,
 } ) => {
 	const [ uploadFile, setFile ] = useState( null );
 	const [ isUploadingInProgress, setIsUploadingInProgress ] = useState( false );
 	const { createErrorNotice } = useDispatch( noticesStore );
+
+	const startUpload = file => {
+		setFile( file );
+		setIsUploadingInProgress( true );
+		onStartUpload( file );
+	};
+
+	// Start the upload process when a file to upload is set.
+	useEffect( () => {
+		if ( ! fileToUpload ) {
+			return;
+		}
+
+		startUpload( fileToUpload );
+	}, [ fileToUpload ] );
 
 	const onSelectURL = useCallback(
 		( videoSource, id ) => {
@@ -42,13 +61,11 @@ const VideoPressUploader = ( {
 
 	const onSelectVideo = useCallback(
 		media => {
-			const isUploadingFile =
-				media?.url && isURL( media?.url ) && getProtocol( media?.url ) === 'file:' && media?.type;
+			const isUploadingFile = isLocalFile( media?.url ) && media?.type;
 
 			// Upload local file.
 			if ( isUploadingFile ) {
-				setFile( media );
-				setIsUploadingInProgress( true );
+				startUpload( media );
 				return;
 			}
 
@@ -66,6 +83,9 @@ const VideoPressUploader = ( {
 
 	const onResetUpload = useCallback( () => {
 		setIsUploadingInProgress( false );
+		if ( isReplacing ) {
+			onReplaceCancel();
+		}
 	}, [] );
 
 	if ( isUploadingInProgress ) {