+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## 0.31.2 - 2023-03-29
+### Changed
+- Minor internal updates.
+
+## 0.31.1 - 2023-03-28
+### Changed
+- Minor internal updates.
+
+## 0.31.0 - 2023-03-27
+### Changed
+- Added option for CUT component to have a tooltip [#29609]
+
+## 0.30.2 - 2023-03-23
+### Changed
+- Updated package dependencies.
+
 ## 0.30.1 - 2023-03-20
 ### Changed
 - RNA: tweak secondary button styling [#29475]