+++ projects/js-packages/webpack-config/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.4.2 - 2023-03-29
+### Changed
+- Minor internal updates.
+
+## 1.4.1 - 2023-03-28
+### Changed
+- Minor internal updates.
+
+## 1.4.0 - 2023-03-27
+### Added
+- Adds compatibility with @svgr/webpack library for loading svg files as react components [#29544]
+
+## 1.3.27 - 2023-03-23
+### Changed
+- Updated package dependencies.
+
 ## 1.3.26 - 2023-03-20
 ### Changed
 - Updated package dependencies. [#29471]