+++ projects/packages/jitm/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.4] - 2023-03-28
+### Fixed
+- JITM: Fix button overflow when text length is too big [#29749]
+
+## [2.3.3] - 2023-03-27
+### Changed
+- JITM: Update styles [#29709]
+
 ## [2.3.2] - 2023-03-20
 ### Changed
 - Updated package dependencies. [#29471]
@@ -554,6 +562,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update Jetpack to use new JITM package
 
+[2.3.4]: https://github.com/Automattic/jetpack-jitm/compare/v2.3.3...v2.3.4
+[2.3.3]: https://github.com/Automattic/jetpack-jitm/compare/v2.3.2...v2.3.3
 [2.3.2]: https://github.com/Automattic/jetpack-jitm/compare/v2.3.1...v2.3.2
 [2.3.1]: https://github.com/Automattic/jetpack-jitm/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/Automattic/jetpack-jitm/compare/v2.2.42...v2.3.0